requirements.txt
streamlit
torch
opencv-python
easyocr
pandas
scikit-learn
sentence-transformers
Pillow
seaborn